clean up errant extern usage

+++ kernel/device/timer.c
@@ -1,18 +1,19 @@
-/*
- *  This file contains functions for initializing and handling interrupts
- *  from the hardware timer.
- */
-
 #include <lib/barelib.h>
 #include <system/interrupts.h>
 #include <system/syscall.h>
 #include <system/queue.h>
 #include <system/sleep.h>
+#include <device/timer.h>
+
+/*
+ *  This file contains functions for initializing and handling interrupts
+ *  from the hardware timer.
+ */
 
 #define TRAP_TIMER_ENABLE 0xa0
 
 volatile uint64_t* clint_timer_addr  = (uint64_t*)0x2004000;
-const uint32_t timer_interval = 100000;
+const uint64_t timer_interval = 100000;
 
 /*
  * This function is called as part of the bootstrapping sequence
